# dev/openapi.yaml
openapi: 3.0.3
info:
  title: Packing List API
  version: 0.1.0
  description: Backpacking gear management API
servers:
  - url: http://localhost:8000
paths:
...
  /api/v1/items:
    get:
      tags:
        - items
      summary: List all items
      operationId: listItems
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - items
      summary: Create a new item
      operationId: createItem
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemCreate'
      responses:
        '201':
          description: Item created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/items/{itemId}:
    get:
      tags:
        - items
      summary: Get item by ID
      operationId: getItem
      security:
        - bearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Item retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - items
      summary: Update item
      operationId: updateItem
      security:
        - bearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemUpdate'
      responses:
        '200':
          description: Item updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - items
      summary: Delete item
      operationId: deleteItem
      security:
        - bearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Item deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  schemas:
...
    Item:
      allOf:
        - $ref: '#/components/schemas/ItemBase'
    ItemBase:
      type: object
      required:
        - id
        - name
        - type
        - is_active
        - manufacturer_id
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
        is_active:
          type: boolean
        manufacturer_id:
          type: string
          format: uuid
        description:
          type: string
          nullable: true
        source_url:
          type: string
          nullable: true
        value:
          type: number
          nullable: true
        weight_grams:
          type: number
          format: float
          nullable: true
        volume_ml:
          type: number
          format: float
          nullable: true
        default_quantity:
          type: integer
          nullable: true
        default_carry_status:
          type: string
          enum: [packed, worn]
          nullable: true
        is_default:
          type: boolean
          nullable: true
        image_blob:
          type: string
          format: byte
          nullable: true
        image_mime_type:
          type: string
          nullable: true
        image_size_bytes:
          type: integer
          nullable: true
        image_width_px:
          type: integer
          nullable: true
        image_height_px:
          type: integer
          nullable: true
        attributes:
          type: object
          additionalProperties: true
          nullable: true
          description: Dynamic per-item values keyed by item type field definitions
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ItemCreate:
      type: object
      required:
        - name
        - type
        - is_active
        - manufacturer_id
      properties:
        name:
          type: string
        type:
          type: string
        is_active:
          type: boolean
        manufacturer_id:
          type: string
          format: uuid
        description:
          type: string
          nullable: true
        source_url:
          type: string
          nullable: true
        value:
          type: number
          format: float
          nullable: true
        weight_grams:
          type: number
          format: float
          nullable: true
        volume_ml:
          type: number
          format: float
          nullable: true
        default_quantity:
          type: integer
          nullable: true
        default_carry_status:
          type: string
          enum: [packed, worn]
          nullable: true
        is_default:
          type: boolean
          nullable: true
        image_blob:
          type: string
          format: byte
          nullable: true
        image_mime_type:
          type: string
          nullable: true
        image_size_bytes:
          type: integer
          nullable: true
        image_width_px:
          type: integer
          nullable: true
        image_height_px:
          type: integer
          nullable: true
        attributes:
          type: object
          additionalProperties: true
          nullable: true
          description: Dynamic per-item values keyed by item type field definitions
    ItemUpdate:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        is_active:
          type: boolean
        manufacturer_id:
          type: string
          format: uuid
        description:
          type: string
          nullable: true
        source_url:
          type: string
          nullable: true
        value:
          type: number
          format: float
          nullable: true
        weight_grams:
          type: number
          format: float
          nullable: true
        volume_ml:
          type: number
          format: float
          nullable: true
        default_quantity:
          type: integer
          nullable: true
        default_carry_status:
          type: string
          enum: [packed, worn]
          nullable: true
        is_default:
          type: boolean
          nullable: true
        image_blob:
          type: string
          format: byte
          nullable: true
        image_mime_type:
          type: string
          nullable: true
        image_size_bytes:
          type: integer
          nullable: true
        image_width_px:
          type: integer
          nullable: true
        image_height_px:
          type: integer
          nullable: true
        attributes:
          type: object
          additionalProperties: true
          nullable: true
          description: Dynamic per-item values keyed by item type field definitions
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            type: object
            required:
              - error
            properties:
              error:
                type: string
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            required:
              - error
            properties:
              error:
                type: string
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            required:
              - error
            properties:
              error:
                type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT